pyproject.toml
[tool.poetry]
name = "poetry-language"
version = "0.1.0"
description = ""
authors = ["saadmhd07"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
together = "^1.3.3"
openai = "^1.52.0"
streamlit = "^1.39.0"
langchain = "^0.3.4"
langchain-together = "^0.2.0"
langchain-community = "^0.3.3"
gtts = "^2.5.3"
av = "12.0.0"
streamlit-webrtc = "^0.47.9"
pydub = "^0.25.1"
streamlit-audiorec = "^0.1.3"
google-cloud-speech = "^2.28.0"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
ffmpeg-python = "^0.2.0"
python-dotenv = "^1.0.1"
flask = "^3.1.0"
flask-cors = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
